feat: Listening Lifestyle — Your Week, After Dark, Year's Traditions
Extracts the three offline-viable features from the listening-lifestyle-hub
branch. Dining (Around the Table) and In Transit (On the Move) are dropped
because the Foursquare export does not include venue category data; see
issue #93 for future approaches.

Retained:
- Your Week: home/away × weekday/weekend play counts with top-10 artist and
  album tabs per day of week
- After Dark: late-night (midnight–4 AM) rate, top artists, hour-by-hour
  breakdown with per-hour tabs
- Year's Traditions: holiday patterns from assumptions file, built-in
  Halloween and Thanksgiving seasons, year-over-year chart, per-holiday tabs

Persona banner updated to 4 badges (Night Owl, Weekend Listener, Globe
Trotter, Holiday Traditionalist) with 3 summary metrics.

Transit/dining tests migrated from test_listening_lifestyle to
test_analysis_utils where they belong; coverage at 70.69%.
